+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Don't write out a journal file if no prices were successfully fetched.
 * Log API errors to `stderr` instead of `stdout`.
 * Improve error messages when the AlphaVantage API returns a
   rate-limit-exceeded error.